+++ flutter/test/sentry_navigator_observer_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -43,7 +45,7 @@ void main() {
   });
 
   group('NativeFrames', () {
-    test('transaction start begins frames collection', () {
+    test('transaction start begins frames collection', () async {
       final currentRoute = route(RouteSettings(name: 'Current Route'));
       final mockHub = _MockHub();
       final native = SentryNative();
@@ -57,7 +59,10 @@ void main() {
 
       sut.didPush(currentRoute, null);
 
-      expect(mockNativeChannel.numberOfBeginNativeFramesCalls, 1);
+      // Handle internal async method calls.
+      await Future.delayed(const Duration(milliseconds: 10), () {
+        expect(mockNativeChannel.numberOfBeginNativeFramesCalls, 1);
+      });
     });
 
     test('transaction finish adds native frames to tracer', () async {
@@ -752,8 +757,8 @@ class Fixture {
 class _MockHub extends MockHub {
   final Scope scope = Scope(SentryOptions(dsn: fakeDsn));
   @override
-  void configureScope(ScopeCallback? callback) {
-    callback?.call(scope);
+  FutureOr<void> configureScope(ScopeCallback? callback) async {
+    await callback?.call(scope);
   }
 }