+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - Bump Android SDK to v6.2.1 ([#945](https://github.com/getsentry/sentry-dart/pull/945))
   - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#621)
   - [diff](https://github.com/getsentry/sentry-java/compare/6.1.4...6.2.1)
+- Bump Cocoa SDK to v7.21.0 ([#947](https://github.com/getsentry/sentry-dart/pull/947))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/master/CHANGELOG.md#7210)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/7.19.0...7.21.0)
 
 ## 6.6.3